requirements.txt
# Social Sentiment & Trend Monitor 프로젝트 의존성 패키지

# 웹 프레임워크
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.25.0
python-multipart>=0.0.6

# 데이터 처리
numpy>=1.24.0
pandas>=2.0.0

# 자연어 처리
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99
openai>=1.0.0

# 데이터 수집
youtube-dl>=2021.12.17
tweepy>=4.14.0
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10

# 시계열 분석
scikit-learn>=1.3.0
scipy>=1.11.0
statsmodels>=0.14.0

# 시각화
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# 데이터베이스
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# 설정 관리
pyyaml>=6.0

# 유틸리티
tqdm>=4.65.0
python-dotenv>=1.0.0
schedule>=1.2.0
wordcloud>=1.9.2

